Moved intrinsic tag parsing to the intrinsics package
