+++ goformation.go
@@ -2,62 +2,18 @@ package goformation
 
 import (
 	"encoding/json"
-	"fmt"
 	"io/ioutil"
 	"strings"
 
 	"github.com/awslabs/goformation/cloudformation"
 	"github.com/awslabs/goformation/intrinsics"
-
-	yamlwrapper "github.com/ghodss/yaml"
-
-	"reflect"
-	"gopkg.in/yaml.v2"
 )
 
 //go:generate generate/generate.sh
 
-type tagUnmarshalerType struct {
-}
-
-func (t *tagUnmarshalerType) UnmarshalYAMLTag(tag string, fieldValue reflect.Value) reflect.Value {
-
-	prefix := "Fn::"
-	if tag == "Ref" || tag == "Condition" {
-		prefix = ""
-	}
-
-	tag = prefix + tag
-
-	output := reflect.ValueOf(make(map[string]interface{}))
-	key := reflect.ValueOf(tag)
-
-	output.SetMapIndex(key, fieldValue)
-
-	return output
-}
-var tagUnmarshaller = &tagUnmarshalerType{}
-var allTags = []string{"Ref", "GetAtt", "Base64", "FindInMap", "GetAZs",
-	"ImportValue", "Join", "Select", "Split", "Sub",
-}
-
-func registerTagMarshallers() {
-	for _, tag := range allTags {
-		yaml.RegisterTagUnmarshaler("!"+tag, tagUnmarshaller)
-	}
-}
-
-func unregisterTagMarshallers() {
-	for _, tag := range allTags {
-		yaml.RegisterTagUnmarshaler("!"+tag, tagUnmarshaller)
-	}
-}
-
 // Open and parse a AWS CloudFormation template from file.
 // Works with either JSON or YAML formatted templates.
 func Open(filename string) (*cloudformation.Template, error) {
-
-	registerTagMarshallers()
 
 	data, err := ioutil.ReadFile(filename)
 	if err != nil {
@@ -65,28 +21,42 @@ func Open(filename string) (*cloudformation.Template, error) {
 	}
 
 	if strings.HasSuffix(filename, ".yaml") || strings.HasSuffix(filename, ".yml") {
-		data, err = yamlwrapper.YAMLToJSON(data)
-		if err != nil {
-			return nil, fmt.Errorf("invalid YAML template: %s", err)
-		}
-
+		return ParseYAML(data)
 	}
 
-	return Parse(data)
+	return ParseJSON(data)
 
 }
 
-// Parse an AWS CloudFormation template (expects a []byte of valid JSON)
-func Parse(data []byte) (*cloudformation.Template, error) {
-
+// ParseYAML an AWS CloudFormation template (expects a []byte of valid YAML)
+func ParseYAML(data []byte) (*cloudformation.Template, error) {
 	// Process all AWS CloudFormation intrinsic functions (e.g. Fn::Join)
-	intrinsified, err := intrinsics.Process(data, nil)
+	intrinsified, err := intrinsics.ProcessYAML(data, nil)
 	if err != nil {
 		return nil, err
 	}
 
+	return unmarshal(intrinsified)
+
+}
+
+// ParseJSON an AWS CloudFormation template (expects a []byte of valid JSON)
+func ParseJSON(data []byte) (*cloudformation.Template, error) {
+
+	// Process all AWS CloudFormation intrinsic functions (e.g. Fn::Join)
+	intrinsified, err := intrinsics.ProcessJSON(data, nil)
+	if err != nil {
+		return nil, err
+	}
+
+	return unmarshal(intrinsified)
+
+}
+
+func unmarshal(data []byte) (*cloudformation.Template, error) {
+
 	template := &cloudformation.Template{}
-	if err := json.Unmarshal(intrinsified, template); err != nil {
+	if err := json.Unmarshal(data, template); err != nil {
 		return nil, err
 	}
